PStraming server/requirements.txt
absl-py==1.3.0
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
bidict==0.22.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==4.4.2
Django==4.1
docx2pdf==0.1.8
executing==1.2.0
Flask==2.2.3
Flask-SocketIO==5.3.2
flatbuffers==23.3.3
fonttools==4.39.3
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.51.0
humanfriendly==10.0
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
imutils==0.5.4
ipykernel==6.21.1
ipython==8.9.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.2
jupyter_core==5.2.0
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
moviepy==1.0.3
mpmath==1.3.0
mysql-connector-python==8.0.31
nest-asyncio==1.5.6
numpy==1.23.4
oauthlib==3.2.2
onnxruntime==1.14.1
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.6.2
proglog==0.1.10
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-docx==0.8.11
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.5
pywin32==305
pyzmq==25.0.0
reportlab==3.6.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch-tb-profiler==0.4.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
tzdata==2022.1
urllib3==1.26.12
wcwidth==0.2.6
Werkzeug==2.2.2
